pyproject.toml
[project]
name = "know-it-all"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "flask-migrate>=4.0.7",
    "lark-oapi>=1.4.19",
    "redis>=5.0.0",
    "pymysql>=1.1.0",
]
